+++ src/bbt-scenarios-readers-adoc_reader.adb
@@ -2,7 +2,7 @@
 -- bbt, the black box tester (https://github.com/LionelDraghi/bbt)
 -- Author : Lionel Draghi
 -- SPDX-License-Identifier: APSL-2.0
--- SPDX-FileCopyrightText: 2024, Lionel Draghi
+-- SPDX-FileCopyrightText: 2025, Lionel Draghi
 -- -----------------------------------------------------------------------------
 
 with Ada.Characters.Latin_1;
@@ -13,11 +13,11 @@ with Ada.Strings.Maps;  use Ada.Strings.Maps;
 
 with GNAT.Regexp;
 
-package body BBT.Scenarios.Readers.MDG_Reader is
+package body BBT.Scenarios.Readers.Adoc_Reader is
 
    -- --------------------------------------------------------------------------
-   Processor       : aliased MDG_Reader;
-   Regexp          : constant String := ("*.md");
+   Processor       : aliased Adoc_Reader;
+   Regexp          : constant String := ("*.adoc");
    Compiled_Regexp : constant GNAT.Regexp.Regexp
      := GNAT.Regexp.Compile (Pattern        => Regexp,
                              Glob           => True,
@@ -25,25 +25,24 @@ package body BBT.Scenarios.Readers.MDG_Reader is
 
    -- --------------------------------------------------------------------------
    Blanks  : constant Character_Set := To_Set (" " & Ada.Characters.Latin_1.HT);
-   -- cf. https://spec.commonmark.org/0.31.2/#blank-line
    Separators : constant Character_Set := To_Set (':') or Blanks;
-   Decorators : constant Character_Set := To_Set ('#'); -- or Markdown_Emphasis
-   -- All surrounding chars that should be trimmed
+   Decorators : constant Character_Set := To_Set ('#');
+   -- AsciiDoc highlight : https://docs.asciidoctor.org/asciidoc/latest/text/highlight/
 
    -- --------------------------------------------------------------------------
    procedure Initialize is
    begin
       Register (Reader     => Processor'Access,
-                For_Format => MDG);
+                For_Format => Adoc);
    end Initialize;
 
    -- --------------------------------------------------------------------------
    function File_Pattern
-     (Reader : MDG_Reader) return String is (Regexp);
+     (Reader : Adoc_Reader) return String is (Regexp);
 
    -- --------------------------------------------------------------------------
    function Is_Of_The_Format
-     (Reader    : MDG_Reader;
+     (Reader    : Adoc_Reader;
       File_Name : String) return Boolean is
    begin
       return GNAT.Regexp.Match (S => File_Name,
@@ -52,7 +51,7 @@ package body BBT.Scenarios.Readers.MDG_Reader is
 
    -- --------------------------------------------------------------------------
    function Find_Heading_Mark
-     (Reader      : MDG_Reader;
+     (Reader      : Adoc_Reader;
       Line        : String;
       First       : out Natural;
       Last        : out Natural;
@@ -60,30 +59,32 @@ package body BBT.Scenarios.Readers.MDG_Reader is
       Title_Last  : out Natural;
       Location    : Location_Type) return Boolean is
 
-   -- In
-   -- ### **Header** : xyz
+   -- === **Header** : xyz
    -- First will point 'H'
    -- Last  will point 'r'
    -- Title_First will point 'x'
    -- Title_Last  will point 'z'
    --
-   -- Refer to https://spec.commonmark.org/0.31.2/#atx-heading
+   -- Refer to https://docs.asciidoctor.org/asciidoc/latest/sections/titles-and-levels/
    -- for specification
+
+   -- Fixme : This code is very similar to Markdown, and should be factorized.
+
    begin
-      -- First point to the first  #
+      -- First point to the first =
       First := Index_Non_Blank (Source => Line, From => Line'First);
 
-      if Line (First) /= '#' then
+      if Line (First) /= '=' then
          return False;
       end if;
 
-      -- let's jump over all others '#'
+      -- let's jump over all others '='
       First := Index (Source => Line,
-                      Set    => Ada.Strings.Maps.To_Set ("#"),
+                      Set    => Ada.Strings.Maps.To_Set ("="),
                       Test   => Ada.Strings.Outside,
                       From   => First + 1,
                       Going  => Forward);
-      -- Now First point to the first character after all '#'
+      -- Now First point to the first character after all '='
       -- First is set, and should not be overwritten from now on
 
       if Line'Last = First then
@@ -91,17 +92,15 @@ package body BBT.Scenarios.Readers.MDG_Reader is
       end if;
 
       if Is_In (Line (First), Blanks) then
-         -- let's jump over blanks
+         -- Let's jump over blanks
          First := Index (Source => Line,
                          Set    => Blanks,
                          Test   => Ada.Strings.Outside,
                          From   => First,
                          Going  => Forward);
       else
-         IO.Put_Warning ("Markdown expect space in Headings after last '#'", Location);
-         --  IO.Put_Warning ("First = " & First'Image);
-         --  IO.Put_Warning ("Line  = " & Line);
-         --  IO.Put_Warning ("Line (First) = " & """" & Line (First)'Image & """");
+         IO.Put_Warning ("Asciidoc expect space in Headings after last '='", Location);
+         -- Couldn't find a ref in asciidoc
       end if;
 
       Find_Token (Source => Line,
@@ -114,7 +113,7 @@ package body BBT.Scenarios.Readers.MDG_Reader is
       -- Last is set, and should not be overwritten from now on
 
       if Line'Last = Last then
-         -- There is noting after "# Scenario" (for example)
+         -- There is noting after "= Scenario" (for example)
          Title_First := 1;
          Title_Last  := 0;
          return True;
@@ -127,8 +126,8 @@ package body BBT.Scenarios.Readers.MDG_Reader is
                             From   => Last + 1,
                             Going  => Forward);
       if Title_First = 0 then
-         -- There is only Separators or Decorators after "# Scenario" (for example)
-         -- Put_Line ("only Separators after # Scenario");
+         -- There is only Separators or Decorators after "= Scenario" (for example)
+         -- Put_Line ("only Separators after = Scenario");
          Title_First := 1;
          Title_Last  := 0;
          return True;
@@ -144,50 +143,24 @@ package body BBT.Scenarios.Readers.MDG_Reader is
          Title_Last := Line'Last;
       end if;
 
+      -- Put_Line ("Token >" & Line (First .. Last) & "<");
+
       return True;
 
    end Find_Heading_Mark;
 
    -- --------------------------------------------------------------------------
-   subtype Marker_String is String (1 .. 3);
-   Tilda_Fence_Marker    : constant Marker_String := 3 * '~';
-   Backtick_Fence_Marker : constant Marker_String := 3 * '`';
-   No_Marker             : constant Marker_String := 3 * ' ';
-   Current_Fence_Marker  : Marker_String := No_Marker;
-   -- Stores what marker was used to open the block
-   -- This allows to ignore code block inserted with the other marker
+   subtype Marker_String is String (1 .. 4);
+   Code_Block_Delimiter : constant Marker_String := 4 * '-';
 
-   function Code_Fence_Line (Reader           : MDG_Reader;
+   function Code_Fence_Line (Reader           : Adoc_Reader;
                              Line             : String;
                              Look_For_Closing : Boolean) return Boolean is
+   -- We use the source block syntax
+   -- https://docs.asciidoctor.org/asciidoc/latest/verbatim/source-blocks/
    begin
-      if Look_For_Closing then
-         -- Looking for the closing marker
-         -- The closing marker must be equal to the opening one
-         if Index (Trim (Line, Left),
-                   Current_Fence_Marker) = 1
-         then
-            -- this is the closing marker, let's reset
-            Current_Fence_Marker := No_Marker;
-            return True;
-         else
-            return False;
-         end if;
-
-      else
-         -- The opening marker is one of the two possibility
-         if Index (Trim (Line, Left), Tilda_Fence_Marker) = 1 then
-            Current_Fence_Marker := Tilda_Fence_Marker;
-            return True;
-         elsif Index (Trim (Line, Left), Backtick_Fence_Marker) = 1 then
-            Current_Fence_Marker := Backtick_Fence_Marker;
-            return True;
-         else
-            return False;
-         end if;
-
-      end if;
-
+      return Index (Trim (Line, Left),
+                    Code_Block_Delimiter) = 1;
    end Code_Fence_Line;
 
-end BBT.Scenarios.Readers.MDG_Reader;
+end BBT.Scenarios.Readers.Adoc_Reader;